+++ clients/wallet/src/binance_sdk_wallet/rest_api/api/asset_api.py
@@ -584,8 +584,6 @@ class AssetApi:
 
                 Query User Delegation History
 
-        * You need to open Enable Spot & Margin Trading permission for the API Key which requests this endpoint
-
         Weight: 60
 
                 Args: