+++ clients/simple_earn/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 5.1.0 - 2025-01-13
+
+### Changed (1)
+
+- Updated `binance-common` library to version `3.4.0`
+
 ## 5.0.0 - 2025-12-22
 
 ### Changed (2)